perf(ai): 生成中の再レンダーを抑制
- 確定メッセージを memo 化した ChatMessage に切り出し、ストリーミング中に
  履歴全体を毎トークン parseGeneration し直さないようにする
- 末尾追従スクロールを messages.length / status の変化時だけに間引き、
  生成中の smooth スクロール連打によるもたつきを解消
